Rework ProcessorCreationContext API
It only provides access to environment related functionality rather than
SakerFiles and other execution time objects. The classpath processor
creator that uses it should be reworked so it caches the classpath files
in the bundle storage and loads it through the environment.

This would ensure that the opened JAR files are properly loaded, and can
also be overwritten instead of loading them in-place.